+++ quantlib/time/api.py
@@ -9,7 +9,7 @@
 
 from .businessdayconvention import BusinessDayConvention
 from .calendar import (
-    Calendar, TARGET, ModifiedFollowing, Following, ModifiedPreceding,
+    Calendar, ModifiedFollowing, Following, ModifiedPreceding,
     Preceding, Unadjusted, holiday_list
     )
 from .calendar_registry import calendar_from_name
@@ -21,6 +21,7 @@ from .calendars.united_states import UnitedStates
 from .calendars.canada import Canada
 from .calendars.switzerland import Switzerland
 from .calendars.japan import Japan
+from .calendars.target import TARGET
 
 from .daycounter import Actual360, Actual365Fixed, DayCounter
 from .daycounters.thirty360 import Thirty360
