+++ quantlib/time/calendars/_target.pxd
@@ -1,0 +1,5 @@
+from quantlib.time._calendar cimport Calendar
+
+cdef extern from 'ql/time/calendars/target.hpp' namespace 'QuantLib':
+    cdef cppclass TARGET(Calendar):
+        TARGET()
