+++ quantlib/time/daycounter.pyx
@@ -4,6 +4,7 @@ from libcpp.string cimport string
 cimport _daycounter
 cimport _date
 cimport _calendar
+cimport quantlib.time.calendars._target as _tg
 
 from date cimport Date
 from calendar cimport Calendar
@@ -145,7 +146,7 @@ cdef class Business252(DayCounter):
     def __cinit__(self, *args, calendar=None):
         cdef _calendar.Calendar* cl
         if calendar is None:
-           cl = <_calendar.Calendar*> new _calendar.TARGET()
+           cl = new _tg.TARGET()
         else:
            cl = (<Calendar>calendar)._thisptr
         self._thisptr = <_daycounter.DayCounter*> new _daycounter.Business252(deref(cl))
@@ -159,4 +160,3 @@ cdef class SimpleDayCounter(DayCounter):
 
     def __cinit__(self, *args):
         self._thisptr = <_daycounter.DayCounter*> new _daycounter.SimpleDayCounter()
-
